Entered inputs is managed by useRef and Handlers
