+++ src/components/Login.jsx
@@ -1,31 +1,18 @@
-import { useState } from "react";
-
+import { useRef } from "react";
 export default function Login() {
-  // const [enteredEmail, setEnteredEmail] = useState("");
-  // const [enteredPassword, setEnteredPassword] = useState("");
-  const [enteredValues, setEnteredValues] = useState({
-    email: "",
-    password: "",
-  });
+  const email = useRef();
+  const password = useRef();
 
   function handleSubmit(event) {
     event.preventDefault();
-    console.log(enteredValues);
-  }
-  function handleInputChange(event) {
-    const { name, value } = event.target;
-    setEnteredValues((prevValues) => ({
-      ...prevValues,
-      [name]: value,
-    }));
-  }
-  // function handleEmailChange(event) {
-  //   setEnteredEmail(event.target.value);
-  // }
 
-  // function handlePasswordChange(event) {
-  //   setEnteredPassword(event.target.value);
-  // }
+    const enteredEmail = email.current.value;
+    const enteredPassword = password.current.value;
+
+    console.log(enteredEmail, enteredPassword);
+
+    //email.current.value = ""; not recommended
+  }
 
   return (
     <form onSubmit={handleSubmit}>
@@ -34,24 +21,12 @@ export default function Login() {
       <div className="control-row">
         <div className="control no-margin">
           <label htmlFor="email">Email</label>
-          <input
-            id="email"
-            type="email"
-            name="email"
-            onChange={(event) => handleInputChange(event)}
-            value={enteredValues.email}
-          />
+          <input id="email" type="email" name="email" ref={email} />
         </div>
 
         <div className="control no-margin">
           <label htmlFor="password">Password</label>
-          <input
-            id="password"
-            type="password"
-            name="password"
-            onChange={(event) => handleInputChange(event)}
-            value={enteredValues.password}
-          />
+          <input id="password" type="password" name="password" ref={password} />
         </div>
       </div>
